backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
openai==1.58.1
langchain==0.3.13
langchain-openai==0.3.0
langchain-text-splitters==0.3.3
python-multipart==0.0.20
pdfplumber==0.11.4
python-docx==1.1.2
pydantic-settings==2.7.1
python-dotenv==1.0.1
